+++ akshare/__init__.py
@@ -2466,9 +2466,10 @@ amac_manager_cancelled_info # 中国证券投资基金业协会-信息公示-诚
 1.10.49 fix: fix stock_hot_rank_em interface
 1.10.50 fix: fix car_energy_sale_cpca interface
 1.10.51 fix: fix fund_money_rank_em interface
+1.10.52 fix: fix stock_financial_hk_report_em interface
 """
 
-__version__ = "1.10.51"
+__version__ = "1.10.52"
 __author__ = "AKFamily"
 
 import sys