+++ docs/changelog.md
@@ -63,6 +63,12 @@
 
 ## 更新说明详情
 
+1.10.52 fix: fix stock_financial_hk_report_em interface
+
+    1. 修复 stock_financial_hk_report_em 接口
+    2. 修复 stock_financial_hk_analysis_indicator_em 接口
+    3. 修复 index_value_name_funddb 接口
+
 1.10.51 fix: fix fund_money_rank_em interface
 
     1. 修复 fund_money_rank_em 接口
@@ -2614,6 +2620,8 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.10.52 fix: fix stock_financial_hk_report_em interface
 
 1.10.51 fix: fix fund_money_rank_em interface